# auth-service/src/main/java/com/sms/auth/dto/PhotoUploadResponse.java
package com.sms.auth.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.time.LocalDateTime;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class PhotoUploadResponse {
    private String photoUrl;
    private LocalDateTime uploadedAt;
}