# auth-service/src/main/java/com/sms/auth/dto/RefreshTokenResponse.java
package com.sms.auth.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class RefreshTokenResponse {
    private String accessToken;
    private String refreshToken;
    private long expiresIn; // in seconds
}