# specs/002-jwt-auth/contracts/api-contracts.yaml
openapi: 3.0.3
info:
  title: JWT Authentication & Profile Management API
  description: |
    API endpoints for JWT token lifecycle management (refresh, logout) and user profile operations.
    Extends the existing auth-service with refresh token support, profile CRUD, password management,
    language preferences, and profile photo uploads.

    **API Standards**: All endpoints return responses in the format `{errorCode: ErrorCode, data: T}`.
    Backend returns only error codes; frontend handles all internationalization.
  version: 1.0.0
  contact:
    name: SMS Development Team
    email: dev@sms.example.com

servers:
  - url: http://localhost:8080
    description: Local development (auth-service)
  - url: http://localhost:9090
    description: API Gateway (routes to auth-service)

tags:
  - name: Authentication
    description: Token lifecycle management (refresh, logout)
  - name: Profile
    description: User profile operations (view, update, photo upload)

paths:
  /api/auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: |
        Exchange a valid refresh token for a new access token (24h) and refresh token (30d).
        Implements token rotation - the provided refresh token becomes invalid after use.

        **Security**: If a refresh token is reused (replay attack), all user sessions are invalidated.
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
            example:
              refreshToken: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_RefreshTokenResponse'
              example:
                errorCode: "SUCCESS"
                data:
                  accessToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  refreshToken: "7c9e6679-7425-40de-944b-e07fc1f90ae7"
                  expiresIn: 86400
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Token replay detected - all sessions invalidated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Error'
              example:
                errorCode: "TOKEN_REPLAY_DETECTED"
                data: null

  /api/auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout and invalidate tokens
      description: |
        Invalidates the current access token (via JWT ID) and all refresh tokens for the user.
        Logs the user out from all devices.
      operationId: logout
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logged out successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_EmptyData'
              example:
                errorCode: "SUCCESS"
                data: null
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /api/auth/me:
    get:
      tags:
        - Profile
      summary: Get current user profile
      description: |
        Returns the authenticated user's profile information including name, email, phone,
        language preference, and profile photo URL.
      operationId: getCurrentUserProfile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ProfileResponse'
              example:
                errorCode: "SUCCESS"
                data:
                  id: "7c9e6679-7425-40de-944b-e07fc1f90ae7"
                  email: "teacher@school.edu.kh"
                  phoneNumber: "+85512345678"
                  name: "Sok Dara"
                  preferredLanguage: "km"
                  profilePhotoUrl: "/uploads/profile-photos/7c9e6679-7425-40de-944b-e07fc1f90ae7/profile.jpg"
                  profilePhotoUploadedAt: "2025-11-15T10:30:00Z"
                  accountStatus: "active"
                  createdAt: "2025-11-01T08:00:00Z"
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /api/users/me:
    put:
      tags:
        - Profile
      summary: Update user profile
      description: |
        Update the authenticated user's profile information (name, phone, language preference).
        Phone number must be unique across all users.
      operationId: updateProfile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
            example:
              name: "Sok Dara"
              phoneNumber: "+85512345678"
              preferredLanguage: "km"
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ProfileResponse'
              example:
                errorCode: "SUCCESS"
                data:
                  id: "7c9e6679-7425-40de-944b-e07fc1f90ae7"
                  email: "teacher@school.edu.kh"
                  phoneNumber: "+85512345678"
                  name: "Sok Dara"
                  preferredLanguage: "km"
                  profilePhotoUrl: "/uploads/profile-photos/7c9e6679-7425-40de-944b-e07fc1f90ae7/profile.jpg"
                  accountStatus: "active"
                  createdAt: "2025-11-01T08:00:00Z"
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Error'
              examples:
                invalidPhone:
                  value:
                    errorCode: "INVALID_PHONE_FORMAT"
                    data: null
                duplicatePhone:
                  value:
                    errorCode: "DUPLICATE_PHONE"
                    data: null
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /api/users/me/password:
    put:
      tags:
        - Profile
      summary: Change password
      description: |
        Change the authenticated user's password. Requires current password verification.
        All sessions except the current one are invalidated after successful password change.
      operationId: changePassword
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
            example:
              currentPassword: "OldPassword123!"
              newPassword: "NewPassword456@"
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_EmptyData'
              example:
                errorCode: "SUCCESS"
                data: null
        '400':
          description: Password validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Error'
              examples:
                incorrectCurrent:
                  value:
                    errorCode: "INCORRECT_PASSWORD"
                    data: null
                weakPassword:
                  value:
                    errorCode: "WEAK_PASSWORD"
                    data: null
                passwordTooShort:
                  value:
                    errorCode: "PASSWORD_TOO_SHORT"
                    data: null
                passwordMissingUppercase:
                  value:
                    errorCode: "PASSWORD_MISSING_UPPERCASE"
                    data: null
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /api/users/me/photo:
    post:
      tags:
        - Profile
      summary: Upload profile photo
      description: |
        Upload a profile photo (JPG or PNG, max 5MB). Replaces existing photo if present.
        Image is validated for content type and sanitized before storage.
      operationId: uploadProfilePhoto
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                photo:
                  type: string
                  format: binary
                  description: Profile photo file (JPG or PNG, max 5MB)
              required:
                - photo
      responses:
        '200':
          description: Photo uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_PhotoUploadResponse'
              example:
                errorCode: "SUCCESS"
                data:
                  photoUrl: "/uploads/profile-photos/7c9e6679-7425-40de-944b-e07fc1f90ae7/profile.jpg"
                  uploadedAt: "2025-11-20T14:25:00Z"
        '400':
          description: Photo validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Error'
              examples:
                sizeExceeded:
                  value:
                    errorCode: "PHOTO_SIZE_EXCEEDED"
                    data: null
                invalidFormat:
                  value:
                    errorCode: "INVALID_PHOTO_FORMAT"
                    data: null
                corruptedImage:
                  value:
                    errorCode: "CORRUPTED_IMAGE"
                    data: null
        '401':
          $ref: '#/components/responses/UnauthorizedError'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token obtained from /api/auth/login or /api/auth/refresh.
        Include in Authorization header: `Bearer {token}`

  schemas:
    ApiResponse_RefreshTokenResponse:
      type: object
      required:
        - errorCode
        - data
      properties:
        errorCode:
          type: string
          description: Error code for client-side i18n lookup, "SUCCESS" for successful operations
          example: "SUCCESS"
        data:
          $ref: '#/components/schemas/RefreshTokenResponse'

    ApiResponse_ProfileResponse:
      type: object
      required:
        - errorCode
        - data
      properties:
        errorCode:
          type: string
          description: Error code for client-side i18n lookup, "SUCCESS" for successful operations
          example: "SUCCESS"
        data:
          $ref: '#/components/schemas/ProfileResponse'

    ApiResponse_PhotoUploadResponse:
      type: object
      required:
        - errorCode
        - data
      properties:
        errorCode:
          type: string
          description: Error code for client-side i18n lookup, "SUCCESS" for successful operations
          example: "SUCCESS"
        data:
          $ref: '#/components/schemas/PhotoUploadResponse'

    ApiResponse_EmptyData:
      type: object
      required:
        - errorCode
        - data
      properties:
        errorCode:
          type: string
          description: Error code for client-side i18n lookup, "SUCCESS" for successful operations
          example: "SUCCESS"
        data:
          type: object
          nullable: true
          description: Null for operations without response payload
          example: null

    ApiResponse_Error:
      type: object
      required:
        - errorCode
        - data
      properties:
        errorCode:
          type: string
          description: Machine-readable error code for client-side i18n lookup
          example: "INVALID_TOKEN"
        data:
          type: object
          nullable: true
          description: Always null for error responses
          example: null

    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          format: uuid
          description: Refresh token UUID obtained from login or previous refresh
          example: "550e8400-e29b-41d4-a716-446655440000"

    RefreshTokenResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: New JWT access token (24h validity)
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refreshToken:
          type: string
          format: uuid
          description: New refresh token UUID (30d validity)
          example: "7c9e6679-7425-40de-944b-e07fc1f90ae7"
        expiresIn:
          type: integer
          description: Access token expiration in seconds
          example: 86400

    ProfileResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: User unique identifier
        email:
          type: string
          format: email
          description: User email address
        phoneNumber:
          type: string
          pattern: '^\+855[1-9]\d{7,8}$'
          description: Cambodia phone number
        name:
          type: string
          nullable: true
          description: User full name (optional)
        preferredLanguage:
          type: string
          enum: [en, km]
          description: User preferred language
        profilePhotoUrl:
          type: string
          nullable: true
          description: Profile photo URL (relative path)
        profilePhotoUploadedAt:
          type: string
          format: date-time
          nullable: true
          description: Photo upload timestamp
        accountStatus:
          type: string
          enum: [active, inactive, locked]
          description: Account status
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp

    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
          nullable: true
          description: User full name (optional)
          example: "Sok Dara"
        phoneNumber:
          type: string
          pattern: '^\+855[1-9]\d{7,8}$'
          description: Cambodia phone number (must be unique)
          example: "+85512345678"
        preferredLanguage:
          type: string
          enum: [en, km]
          description: Preferred interface language
          example: "km"

    ChangePasswordRequest:
      type: object
      required:
        - currentPassword
        - newPassword
      properties:
        currentPassword:
          type: string
          description: Current password (for verification)
          example: "OldPassword123!"
        newPassword:
          type: string
          minLength: 8
          description: |
            New password (must meet strength requirements):
            - At least 8 characters
            - At least 1 uppercase letter
            - At least 1 lowercase letter
            - At least 1 digit
            - At least 1 special character
          example: "NewPassword456@"

    PhotoUploadResponse:
      type: object
      properties:
        photoUrl:
          type: string
          description: Profile photo URL (relative path)
          example: "/uploads/profile-photos/7c9e6679-7425-40de-944b-e07fc1f90ae7/profile.jpg"
        uploadedAt:
          type: string
          format: date-time
          description: Upload timestamp
          example: "2025-11-20T14:25:00Z"

  responses:
    UnauthorizedError:
      description: Authentication required or token invalid/expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse_Error'
          example:
            errorCode: "UNAUTHORIZED"
            data: null

security:
  - bearerAuth: []